This .Rmd document do web scraping from "center watch" website. It contains the functions to return the new drug list proved by FDA for specified year (I have tried 2011 - 2013).
The drug list returns from this .Rmd list contains the drugs categories, name and industry. Please match the drug name to our data set and find the hospital using these drugs. Then we can get a pair of hospital and industry via the new drugs they used. If we find this pair in the money transfer data. There might be a fraud